test: introduce `samevalues()` TAP checker
The introduced `samevalues()` helper checks that values in the range
from 1, to `table.maxn()` of the given table are exactly the same. It
may be useful to test the consistency of JIT and VM behaviour.
Originally, the `arr_is_consistent()` function was introduced in the
<tarantool-tests/gh-6163-min-max.test.lua>. `samevalues()` has the same
functionality (except for the usage of `table.maxn()` instead of the `#`
operator to be sure, that the table we check isn't a sparse array).

Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>
(cherry picked from commit 4fc148553c3656ce219bec572f68f1f067f06dd1)