test: fix fix-mips64-spare-side-exit-patching
This test may be flaky due to a sparse filling of the vector with
traces. Hence, `misc.getmetrics().jit_trace_num` (which is the total
number of traces) is not the last trace number, so there is no mcode for
it.

This patch fixes flakyness from two sides:
* The `fillmcode()` function now starts a search from the given trace
  itself, so there is no possible inconsistency with the growing vector.
* The finding of the last trace number in the aforementioned test is now
  more bulletproof: we scan all numbers from the metrics value to the
  maxtrace value. The last number with an existing mcode is chosen as
  the resulting last trace number.

Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>
(cherry picked from commit 7d5901dda306af9112a7bec8a5acbecc25542acd)