Restore cur_L for specific Lua/C API use case.
Thanks to Peter Cawley.

(cherry-picked from commit e86990f7f24a94b0897061f25a84547fe1108bed)

There is a case of valid Lua C API usage when non-throwing code is
executed on coroutine, other than one that an entry-point C function was
called with, which results in a segmentation fault because cur_L is not
being restored correspondingly. For a comprehensive example, see the
test case added within this patch.

According to the analysis done in LuaJIT/LuaJIT#1066, to hit this issue,
we first need to call into a C function. That C function needs to change
cur_L by calling lua_resume/lua_p?call, and then the only problematic
path is leaving the C function by throwing an error, which is caught
by the fast-function x?pcall. When x?pcall then returns to its caller,
we can be executing bytecode in the VM with an incorrect cur_L, and
things will go badly if we enter a trace before correcting cur_L.

The fix proposed in LuaJIT/LuaJIT#740 was to set cur_L on error throw;
however, the analysis from LuaJIT/LuaJIT#1066 suggests that it can be
set on the catching phase.

This patch changes the error-catching routine, so now the patch sets the
actual cur_L there. Now it is possible to throw errors on non-executing
coroutines, which is a violation of the Lua C API. So, even though it is
now possible, that behavior should be avoided anyway.

Maxim Kokryashkin:
* added the description and the test for the problem

Resolves tarantool/tarantool#6323
Part of tarantool/tarantool#9145

Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Signed-off-by: Igor Munkin <imun@tarantool.org>
(cherry picked from commit 0e6357a1a7e230586747f82209a6737ee1d56c70)