test: introduce tests for debugging extensions
This patch adds tests for LuaJIT debugging extensions for lldb and gdb.
The tests are written in Python's unittest framework [1].

Most of the tests are failed for the lldb due to outdated extension
sources and overcomplicated hard-coded C structures fields
introspection. Hence, tests for LLDB are disabled since they are failing
anyway.

The tarantool-debugger-tests target is introduced. This target is
included in the LuaJIT-check-all target but not in the LuaJIT-test
target to avoid it running for all LuaJIT builds by default in CI.

[1]: https://docs.python.org/3/library/unittest.html

Co-authored-by: Sergey Kaplun <skaplun@tarantool.org>
Reviewed-by: Evgeniy Temirgaleev <e.temirgaleev@tarantool.org>
Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Signed-off-by: Sergey Kaplun <skaplun@tarantool.org>
(cherry picked from commit 9ca7087e2c6e4c95cf3bce330222ee352d31de5a)