Improve last commit.
(cherry-picked from commit 113a168b792cd367822ec04cdc2ef32facd28efa)

The `noconflict` function performs a search for the specified IR
operations applied to the ref. The `noload` parameter specifies
whether an additional load from the same ref should be
considered a conflict too. As one can see, the `noconflict` call
that was added in the previous patch does not require a search
for conflicting loads because another `noconflict` is always
executed beforehand and it checks for conflicting loads.

Maxim Kokryashkin:
* added the description for the patch

Part of tarantool/tarantool#9595