profilers: introduce event reader module
There is no error-checking in profilers which results in raw Lua
errors being reported in cases of non-existing paths or corrupt
file structures. This patch adds a profiler-agnostic module for
event parsing, which is going to be used to introduce
user-friendly errors. It is impossible to enable it right now
because it should be included in Tarantool's build procedure
first.

Part of tarantool/tarantool#9217
Part of tarantool/tarantool#5994

Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Signed-off-by: Sergey Kaplun <skaplun@tarantool.org>
(cherry picked from commit df5a07218dba8cefa370499a1a54ba75fb73b99b)