test: enable <ffi_convert.lua> in LuaJIT-tests
This patch changes `dofile()` (which uses an unreliable relative file
path and mutates `_G` with global functions to be defined) to the
corresponding require and renames the C library in the `require()` call.
It adds groups with the names of subtests and enables the test in
<index>. It renames structure and enum typedefs to avoid conflicts with
cdef in other tests. The GC threshold is increased since all tests run
in a single process, so the number of GC objects is greater. Also, the
one test case is splitted into several.

Part of tarantool/tarantool#9398

Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Signed-off-by: Sergey Kaplun <skaplun@tarantool.org>
(cherry picked from commit a8c69d818642ea18327513b4164c2d282948461d)