test: more cautious usage of LD_PRELOAD for ASan
This patch adds a comment for clarification about the usage of different
versions of ASan implementations for different compilers. Also, it
replaces ' ' with a ':' as a separator in `LD_PRELOAD` to make it more
robust to the CMake list semantics and removes unnecessary double
quotes.

Needed for tarantool/tarantool#9656

Co-authored-by: Sergey Bronnikov <sergeyb@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Signed-off-by: Sergey Kaplun <skaplun@tarantool.org>
(cherry picked from commit d069e8e8e31978e13687977070b6ffa0fdc2bce7)