test: fix lj-802-panic-at-mcode-protfail GCC+ASan
The aforementioned test uses `LD_PRELOAD` to mock the system call
to `mprotect()`. The GCC implementation of libasan requires that
the ASan library go first in the `LD_PRELOAD` list. This patch
tweaks the behaviour to avoid failure of the test. OTOH, this
patch introduces warnings related to the libc leaks for a couple
of tests, but they are not treated as test failures. This should
be fixed by target setting LD_PRELOAD only for necessary tests
when we start using CTest instead of `prove`.

Resolves tarantool/tarantool#9656
Relates to tarantool/tarantool#9898

Co-authored-by: Sergey Bronnikov <sergeyb@tarantool.org>
Reviewed-by: Igor Munkin <imun@tarantool.org>
Reviewed-by: Sergey Kaplun <skaplun@tarantool.org>
Signed-off-by: Sergey Kaplun <skaplun@tarantool.org>
(cherry picked from commit 888adcfb72e361b150a9038407eb1dc9bd3392bd)