test: fix lj-802-panic-at-mcode-protfail GCC+ASan
The aforementioned test uses `LD_PRELOAD` to mock the system call to
`mprotect()`. The GCC implementation of libasan requires that the ASan
library go first in the `LD_PRELOAD` list. This patch tweaks the
behaviour to avoid failure of the test. OTOH, this patch intorduces
warnings related to the libc leaks for a couple of tests, but they are
not treated as test failures. This should be fixed by target setting
LD_PRELOAD only for necessary tests when we start using CTest instead of
`prove`.

Part of tarantool/tarantool#9656

Co-authored-by: Sergey Bronnikov <sergeyb@tarantool.org>