test: enable <misc/hook_norecord.lua> LuaJIT test
This patch moves the aforementioned test from the <misc> to the <trace/>
directory, includes it in <index>, and names the subtest. Also, it
slightly refactors the code to make it closer to our code style.

The test is adjusted to be runable in the test suite by adding the
`jit.flush()` to avoid collisions of traces.

Part of tarantool/tarantool#9398

Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Signed-off-by: Sergey Kaplun <skaplun@tarantool.org>