ARM64: Make tobit conversions match JIT backend behavior.
Thanks to Peter Cawley.

(cherry picked from commit cdc2db3aeac442e139c3b1fda715fdfa78da8791)

On the aarch64 architecture, the `tobit()` function incorrectly
truncates instead of rounding to the nearest integer. This leads to
inconsistency between various VMs (comparing to x86/x64, for example)
and between the arm64 VM and JIT.

This patch uses an addition of 2^52 + 2^51 constant to the given
argument to convert it to the nearest integer value. After reading the
lowest 32 bits of the register, we get the expected result.

Sergey Kaplun:
* added the description and the test for the problem

Part of tarantool/tarantool#12880

Reviewed-by: Evgeniy Temirgaleev <e.temirgaleev@tarantool.org>
Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Signed-off-by: Sergey Kaplun <skaplun@tarantool.org>