test: off JIT for routines in <lang/stackov.lua>
This patch is a follow-up to the commit
6a3a97cd8111a2164143fcf81439f70588e0be42 ("test: enable
<misc/stackovc.lua> LuaJIT test"). The enabled test may raise an error
for the specific builds (see [1], for example) when the stack overflow
error is raised on the trace exit during the restoration from the
snapshot. In that case, LuaJIT VM can't restore the full traceback, and
the corresponding check in the test is failed.

This patch disables JIT for these functions with recursive calls to
avoid such failures.

[1]: https://github.com/tarantool/luajit/actions/runs/10717386746

Follows up tarantool/tarantool#9398

Reviewed-by: Maxim Kokryashkin <m.kokryashkin@tarantool.org>
Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Signed-off-by: Sergey Kaplun <skaplun@tarantool.org>
(cherry picked from commit ff62d0cf991aafc1421dcf9452b6bb5dfeebc9dd)