codehealth: fix warnings for the codespell 2.3.0
This patch fixes the typo 'snapshoting' -> 'snapshotting' in the
<lj-736-BC_UCLO-triggers-infinite-loop.test.lua>. The <lint.yml>
workflow is updated to the corresponding version of the codespell.

Also, it adds `tru` and `fal` to the ignore list since they are used in
the LuaJIT JIT dump output.

Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Signed-off-by: Sergey Kaplun <skaplun@tarantool.org>
(cherry picked from commit 921cc54d323859fe8e22cb374766d9419ed91ee8)