ffi/gc: restore back rehashing of finalizers table
This is a follow-up to the commits
2115828bc6fae911215cd07593df2ae0a83a0f42 ("FFI: Drop finalizer table
rehash after GC cycle.") and bfcbaa70e7b3e720578c05db3f66d832419e8566
("Drop unused function wrapper."). After them, the rehashing of the cdata
finalizer table at the end of the GC cycle is dropped. Without
reshashing of this table, the table increases the estimated amount of
memory for the GC. Hence, with the bigger `estimate`, the threshold
before starting the GC cycle is increased too. This allows allocating
more cdata objects and increasing the size of the finalizer table again.
This increases the memory estimate again and so on. As a result, we have
unlimited memory growth without rehashing of the table for the
cdata-intensive workloads.

This patch reverts back the code changes (but not the test) of the
aforementioned commits. Also, it fixes the possible crash after
rehashing of the cdata finalizers table by adding the protected call to
the GC steps on the trace itself and on the trace exit.

Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Signed-off-by: Sergey Kaplun <skaplun@tarantool.org>